+++ packages/webdriverio/CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 
+## 0.3.2
+
+### Patch Changes
+
+- Updated dependencies [fcd39f8]
+  - @testivai/witness@2.0.2
+
 ## 0.3.1
 
 ### Patch Changes